export component/move to new folder/fix thumb nails
