+++ components/index.css
@@ -44,5 +44,5 @@
 @import 'toggle/index.css';
 @import 'typography/index.css';
 @import 'welcome/index.css';
-
+@import 'photo-gallery-in-page/index.css';
 /*** Add new component below */
