+++ components/photo-gallery-in-page/PhotoGalleryInPage.vue
