+++ components/photo-gallery-in-page/__tests__/PhotoGalleryInPage.test.js
