+++ components/photo-gallery-in-page/index.css
@@ -1,0 +1,104 @@
+.photo-gallery-in-page {
+  color: #fff;
+  background: #000;
+  &__slider {
+    display: flex;
+    align-items: center;
+    margin-bottom: 5px;
+    img {
+      width: 100%;
+      object-fit: contain;
+    }
+
+    figure {
+      cursor: pointer;
+      padding: 0;
+      margin: 0;
+      display: flex;
+    }
+  }
+
+  &__chevron {
+    color: #fff;
+    cursor: pointer;
+  }
+
+  &--container {
+    display: flex;
+    justify-content: center;
+    width: 80%;
+    height: 600px;
+    margin: auto;
+    background: #000;
+    @media (--bp-max-mobile) {
+      width: 100%;
+      height: 400px;
+    }
+  }
+
+  &__thumbnails {
+    display: flex;
+    height: 74px;
+    width: 80%;
+    margin: auto;
+    overflow-x: auto;
+
+    .thumb {
+      cursor: pointer;
+      margin-right: 5px;
+      opacity: .7;
+      width: 100%;
+      flex: 1;
+      &.active {
+        opacity: 1;
+        fill: rgba(153, 162, 170, .25);
+      }
+    }
+
+    img {
+      width: auto;
+      height: 100%;
+    }
+    @media (--bp-max-mobile) {
+      width: 100%;
+    }
+  }
+
+  &__images-count {
+    width: 80%;
+    margin: auto;
+    padding-top: 11px;
+    font-size: 16px;
+    @media (--bp-max-mobile) {
+      width: 90%;
+    }
+  }
+
+  &__title {
+    width: 80%;
+    margin: 18px auto;
+    font-size: 18px;
+    font-weight: 600;
+    line-height: 25px;
+    @media (--bp-max-mobile) {
+      width: 90%;
+    }
+  }
+
+  &__description {
+    width: 80%;
+    margin: auto;
+    font-size: 16px;
+    line-height: 25px;
+    @media (--bp-max-mobile) {
+      width: 90%;
+    }
+  }
+
+  .arrow-wrapper {
+    padding: 10px;
+    @media (--bp-max-mobile) {
+      display: none;
+    }
+  }
+}
