+++ components/photo-gallery-in-page/sample.json
