+++ components/photo-gallery-in-page/stories/photo-gallery-in-page.md
